splitting tests into different files, adding a condition to ValidateSnap
